refactor(wrapper): Update wrapper as module and include conditional initialization
Implement gwrapper as module
Check if repository exists before creating instance
